2023/03/17

+++ datasets/utils.py
@@ -6,7 +6,7 @@ def row_processor(row):
     :param row:
     :return: label: [gt] data: [0:3]
     """
-    label = np.array(row[0], np.int32)  # 64
+    label = np.array(row[0], np.float32)  # 64
     labels = np.reshape(np.eye(4)[label.astype(np.int32)], (64, 4))  # time_slot, cls -> one-hot encoding
     return {"label": labels.astype(np.float32), "data": np.array(row[1], dtype=np.float32)}
